+++ lib/blkid/read.c
@@ -197,7 +197,8 @@ static int parse_dev(blkid_cache cache, blkid_dev *dev, char **cp)
 	start = skip_over_blank(start + 1);
 	end = skip_over_word(start);
 
-	DBG(DEBUG_READ, printf("device should be %*s\n", end - start, start));
+	DBG(DEBUG_READ, printf("device should be %*s\n",
+			       (int)(end - start), start));
 
 	if (**cp == '>')
 		*cp = end;
