+++ resize/resize2fs.c
@@ -186,7 +186,7 @@ errcode_t adjust_fs_info(ext2_filsys fs, ext2_filsys old_fs, blk_t new_size)
 	unsigned long	i, j, old_desc_blocks, max_group;
 	unsigned int	meta_bg, meta_bg_size;
 	int		has_super;
-	__u64		new_inodes;	/* u64 to check for overflow */
+	unsigned long long new_inodes;	/* u64 to check for overflow */
 
 	fs->super->s_blocks_count = new_size;
 
@@ -227,7 +227,7 @@ retry:
 	/*
 	 * Adjust the number of inodes
 	 */
-	new_inodes =(__u64)fs->super->s_inodes_per_group * fs->group_desc_count;
+	new_inodes =(unsigned long long) fs->super->s_inodes_per_group * fs->group_desc_count;
 	if (new_inodes > ~0U) {
 		fprintf(stderr, _("inodes (%llu) must be less than %u"),
 				   new_inodes, ~0U);
@@ -303,7 +303,7 @@ retry:
 			(old_fs->desc_blocks - fs->desc_blocks);
 		if (new < 0)
 			new = 0;
-		if (new > fs->blocksize/4)
+		if (new > (int) fs->blocksize/4)
 			new = fs->blocksize/4;
 		fs->super->s_reserved_gdt_blocks = new;
 		if (new == 0)
